refactor: extract regression metric helpers
